enable sf again, remove compiled geos
this PR enables sf, which with version 0.4-1 works again with GEOS versions older than 3.5.0; hence, I also outcommented the lines that install GEOS from source. This leaves us with a somewhat older GEOS, and some functions in sf that are not available (e.g., st_voronoi). Also, sf::st_makevalid needs a liblwgeom-dev more recent than in this distro, so that is not being picked up when compiling sf, which makes liblwgeom-dev a candidate to be removed from this Dockerfile collection.